+++ src/Pim/Bundle/TransformBundle/Denormalizer/Flat/ProductValue/MediaDenormalizer.php
@@ -10,6 +10,8 @@ use Pim\Bundle\CatalogBundle\Manager\MediaManager;
  * @author    Gildas Quemener <gildas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * TODO: should be deleted
  */
 class MediaDenormalizer extends AbstractValueDenormalizer
 {